pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["discopat"]

[project]
name = "discopat"
version = "0.3.0"
readme = "README.rst"
description = "DISCOver PATterns (DISCOPAT)"
license = {file = "LICENSE"}
authors = [
  {name = "Mansour Benbakoura", email = "mansour.benbakoura@inria.fr"}
]
classifiers = [
  "Intended Audience :: Science/Research",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
]

requires-python = ">=3.8"

dependencies = [
  "filterpy",
  "h5py",
  "imageio",
  "ipykernel",
  "matplotlib",
  "moviepy<2",
  "numpy",
  "osfclient",
  "PyYAML",
  "scikit-image",
  "torch",
  "torchaudio",
  "torchvision",
  "tqdm",
]

[project.optional-dependencies]
dev = [
  "pycocotools",
  "ruff",
]
doc = [
  "sphinx",
  "sphinx-gallery",
  "sphinx_rtd_theme",
]
test = [
  "pytest",
  "pytest-cov",
]

[project.urls]
Repository = "https://github.com/mansour-b/discopat"
Documentation = "https://discopat.readthedocs.io"

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]

line-length = 80
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN204",
  "COM812",
  "EM101",
  "TRY003",
]

fixable = ["ALL"]
unfixable = []

exclude = ["examples"]

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
